BACKEND/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
PyMySQL>=1.1
Flask-Migrate>=4.0
Flask-JWT-Extended>=4.6
python-dotenv>=1.0
huggingface_hub>=0.23
Flask-Cors>=4.0
